only run  uncertainty if it wasn't yet evaluated

+++ systematics/signal/10_TrkRecoEff/runSystematicUncertainty.sh
@@ -1,22 +1,34 @@
 #!bin/bash
 
-ptCut=50
+#ptCut=60
 
 ###########################################################################################################################
 # edit analyzer.cc file and run analysis
-cd removeHits
-sed -e "s/MuonTracksAllHits_pt\[i\] < __ptCut__/MuonTracksAllHits_pt\[i\] < ${ptCut}/" analyzer.cc > analyzerAux.cc
-mv analyzerAux.cc analyzer.cc
-#make clean
-make
-./analyzer filelist_data.txt analyzer_histogram_data.root > ../result_data.log
-./analyzer filelist_mc.txt   analyzer_histogram_mc.root   > ../result_mc.log
-sed -e "s/MuonTracksAllHits_pt\[i\] < ${ptCut}/MuonTracksAllHits_pt[i] < __ptCut__/" analyzer.cc > analyzerAux.cc
-mv analyzerAux.cc analyzer.cc
-cd ..
+
+
+if [ ! -f "logFiles/result_ptCutEq${ptCut}.log" ] 
+then
+    echo ""
+    echo "%%%%%%%%% result_ptCutEq${ptCut}.log does not exists: Estimate trk reco eff uncertainty: "
+    echo ""
+
+    cd removeHits
+    sed -e "s/MuonTracksAllHits_pt\[i\] < __ptCut__/MuonTracksAllHits_pt\[i\] < ${ptCut}/" analyzer.cc > analyzerAux.cc
+    mv analyzerAux.cc analyzer.cc
+    #make clean
+    make
+    echo "before analysis"
+    ./analyzer filelist_data.txt analyzer_histogram_data.root > ../result_data.log
+    ./analyzer filelist_mc.txt   analyzer_histogram_mc.root   > ../result_mc.log
+    echo "after analysis"
+    sed -e "s/MuonTracksAllHits_pt\[i\] < ${ptCut}/MuonTracksAllHits_pt[i] < __ptCut__/" analyzer.cc > analyzerAux.cc
+    mv analyzerAux.cc analyzer.cc
+    cd ..
+    python getSysUncertainty.py > result_ptCutEq${ptCut}.log
+    cat result_ptCutEq${ptCut}.log
+fi
+
 ###########################################################################################################################
 
-python getSysUncertainty.py > result.log
-cat result.log
 
 rm *~
